refactor: Add CI/CD and Unit Test skills to extraSkills component
Update the extraSkills component to include the skills "CI/CD" and "Unit Test" in the list of extra skills. This will provide a more comprehensive overview of the developer's skillset.

Also, refactor the progress bar values in the programmingLanguages component. Increase the progress for Java to 75% and Spring Boot to 60% to better reflect the developer's proficiency in these languages and frameworks.

In addition, update the knowledge component to dynamically render knowledge cards using data from the knowledgeData array in the utils folder. This allows for easier maintenance and scalability of the knowledge section.

Lastly, add the knowledgeData array to the utils folder, which contains information about different areas of expertise in web development, backend development, and frontend development. Each knowledge card includes a title, description, title size, description size, and an associated icon.

These changes improve the accuracy and completeness of the developer's skills and knowledge representation in the portfolio website.